# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API DeGusto",
    "description": "APIs para el sistema de DeGusto",
    "contact": {
      "url": "https://www.linkedin.com/in/guillermo-yeltsin-baltodano-león-86938413b/"
    },
    "version": "1.0"
  },
...
  "servers": [{ "url": "http://localhost:3000/api/" },{"url": "http://localhost:3000/"}],
...
  "paths": {
...
    "/dish/dishCategory":{
        "get":{
            "tags": [ "Dish" ],
            "summary": "Lista de todas las categorías de los platos",
            "description": "Lista de las primeras 5 categorías de los platos más vendidos",
            "operationId": "getDishCategory",
            "parameters": [],
            "responses": {
                "200": {
                    "description": "successful operation",
                    "content": {
                        "application/json":{
                            "schema":{
                                "type": "array",
                                "items":{
                                    "$ref": "#/components/schemas/CategoryDish"
                                }
                            }
                        }
                    }
                },
                "404": {
                    "description": "No implementado"
                }
            }
        }
    },
    "/dish/popular/{dishCategoryId}":{
        "get":{
            "tags": [ "Dish" ],
            "summary": "Lista de los platos más populares",
            "description": "Lista de los 5 platos más vendidos por su categoría",
            "operationId": "getPopularDish",
            "parameters": [
                {
                    "name": "dishCategoryId",
                    "in": "path",
                    "description": "ID de la categoría del plato",
                    "required": true,
                    "schema": {
                      "type": "integer",
                      "format": "int64"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "successful operation",
                    "content": {
                        "application/json":{
                            "schema":{
                                "type": "array",
                                "items":{
                                    "$ref": "#/components/schemas/Dish"
                                }
                            }
                        }
                    }
                },
                "404": {
                    "description": "No implementado"
                }
            }
        }
    },
    "/dish/business/{dishBusinessId}":{
        "get":{
            "tags": [ "Dish" ],
            "summary": "Lista de los platos por su negocio",
            "description": "Lista de todos platos del negocio",
            "operationId": "getDishByBusinessId",
            "parameters": [
                {
                    "name": "dishBusinessId",
                    "in": "path",
                    "description": "ID del negocio del plato",
                    "required": true,
                    "schema": {
                      "type": "integer",
                      "format": "int64"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "successful operation",
                    "content": {
                        "application/json":{
                            "schema":{
                                "type": "array",
                                "items":{
                                    "$ref": "#/components/schemas/Dish"
                                }
                            }
                        }
                    }
                },
                "404": {
                    "description": "No implementado"
                }
            }
        }
    },
...
  },
  "components": {
    "schemas": {
...
        "CategoryDish":{
            "type": "object",
            "properties": {
                "dishCategoryID":{
                    "type": "integer",
                    "format": "int64",
                    "example":0
                },
                "dishCategoryName":{
                    "type": "string",
                    "example":"Bebida"
                }
            }
        },
        "Dish":{
            "type": "object",
            "properties": {
                "dishId": {
                    "type": "integer",
                    "format": "int64",
                    "example": 0
                },
                "dishName":{
                    "type": "string",
                    "example":"Tallarín con pollo"
                },
                "dishDescription":{
                    "type": "string",
                    "example":"Descripción de lo que lleva el plato"
                },
                "dishPrice":{
                    "type": "number",
                    "example":20.00
                },
                "dishPhoto":{
                    "type": "string",
                    "example":"Ruta final donde esta almacenada la imagen"
                },
                "dish_BusinessId": {
                    "type": "integer",
                    "format": "int64",
                    "example": 0
                },
                "dish_CategoriesId": {
                    "type": "integer",
                    "format": "int64",
                    "example": 0
                },
                "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "La imagen a subir"
                }
            }
        }
    }
  }
}